fix(image-gen): move fanOutJobs out of route file
Next.js App Router rejects non-handler exports from route.ts modules
("Route does not match the required types of a Next.js Route" at
build time). Moved fanOutJobs and its private helpers from
app/api/platform/image/ingest/route.ts to lib/image/fan-out.ts.

The route now imports fanOutJobs from the lib; the test does the
same. Local `npm run build` passes.
